back-end/requirements/production.txt
# Production requirements
-r base.txt

# Production server
gunicorn==21.2.0

# Production database (already in base.txt)
# psycopg2-binary==2.9.9

# Static file serving (already in base.txt)  
# whitenoise==6.6.0

# Security
django-csp==3.7
django-permissions-policy==4.19.0

# Monitoring & Logging
sentry-sdk[django]==1.38.0

# Performance
django-redis==5.4.0